feat(announcement): add announcement module with dashboard card and management page feat(inspection): add export-to-excel endpoints refactor(money): migrate amount fields from Integer to BigDecimal perf(report,inspection): batch prefetch to fix N+1 and add id string serialization

+++ apps/edifice-core/src/main/java/com/qsy/edifice/domain/dto/CreateCollectionRecordDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -20,7 +21,7 @@ public class CreateCollectionRecordDto {
     private Long projectStageId;
 
     @Schema(description = "回款金额（元）", requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer amount;
+    private BigDecimal amount;
 
     @Schema(description = "实际回款日期", requiredMode = Schema.RequiredMode.REQUIRED)
     private LocalDate collectDate;
